Fixed compile errors

+++ spec/avrstate.d
@@ -89,7 +89,7 @@ final class AvrState(AvrChipSpec chip) : MachineState {
     ReferenceRegister!(ubyte)[64] ioRegisters;
     private ReferenceRegister!ushort[string] _refregs;
     int resetEEMPECounter = 0;
-    int resetEEPECounter = 0;
+    ulong resetEEPECounter = 0;
 
     @property final size_t programCounter() {
         return instructions.next.address/2;
